fix hero
