+++ src/components/Hero.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { HeroProps } from '../types/HeroTypes';
+import { HeroProps } from '../types/HeroProps';
 import './Hero.css';
 
 const Hero: React.FC<HeroProps> = ({
@@ -10,17 +10,21 @@ const Hero: React.FC<HeroProps> = ({
 }) => {
   return (
     <div
-      className='hero-section'
+      className='hero-section p-5 text-center bg-image shadow-2-strong'
       style={{
         backgroundImage: `url(${backgroundImage})`,
       }}
     >
       <div className='mask'>
-        <div className='text-content'>
-          <h1>{title}</h1>
-          {subtitle && <p>{subtitle}</p>}
+        <div className='text-white'>
+          <h1 className='mb-3'>{title}</h1>
+          {subtitle && <p className='mb-3 mx-5'>{subtitle}</p>}
           {buttons.map((button, index) => (
-            <a key={index} href={button.href} className='hero-button'>
+            <a
+              key={index}
+              href={button.href}
+              className='btn btn-outline-light btn-lg m-1'
+            >
               {button.label}
             </a>
           ))}
