+++ src/index.css
@@ -13,6 +13,10 @@
   -moz-osx-font-smoothing: grayscale;
 }
 
+* {
+  box-sizing: border-box;
+}
+
 a {
   font-weight: 500;
   color: #646cff;
@@ -24,101 +28,4 @@ a:hover {
 
 html {
   font-size: 16px;
-}
-
-body {
-  margin: 0;
-  display: flex;
-  place-items: center;
-  min-width: 320px;
-  min-height: 100vh;
-}
-
-h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
-}
-
-button {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: #1a1a1a;
-  cursor: pointer;
-  transition: border-color 0.25s;
-}
-button:hover {
-  border-color: #646cff;
-}
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-}
-
-@media (prefers-color-scheme: light) {
-  :root {
-    color: #213547;
-    background-color: #ffffff;
-  }
-  a:hover {
-    color: #747bff;
-  }
-  button {
-    background-color: #f9f9f9;
-  }
-}
-
-.top-padding-100 {
-  padding-top: 100px;
-}
-
-.h-full {
-  height: 100vh;
-}
-
-/* Hero */
-#hero {
-  margin-top: 50px;
-  padding-top: 20px;
-  background-image: url('./assets/hero-img.jpg');
-  text-align: center;
-  background-color: gray;
-  color: whitesmoke;
-}
-
-.jumbotron {
-  opacity: 0.9;
-  color: whitesmoke;
-}
-
-/* About Me */
-#about-me {
-  background-color: whitesmoke;
-}
-
-/* Projects */
-#projects h2 {
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-/* Contact Me */
-#contact {
-  text-align: center;
-  background-color: gray;
-  color: whitesmoke;
-}
-#contact h2 {
-  text-align: center;
-}
-
-.sub-heading {
-  font-size: 1.5rem;
-}
-
-#email {
-  font-size: 2rem;
 }
