--- src/types/HeroTypes.ts
@@ -1,9 +1,0 @@
-export interface HeroProps {
-  backgroundImage: string;
-  title: string;
-  subtitle?: string;
-  buttons?: {
-    label: string;
-    href: string;
-  }[];
-}
